backend/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
model-mommy==1.6.0
